requirements.txt
Django==1.7
django-ckeditor==4.4.7
django-taggit==0.12.3
django-watson==1.1.5
